Add end-to-end encryption for inputs
I'm not using AE for this as before since it's complex, but the server has been changed so that it refuses to listen to any input before the user first sends the encrypted zero block for verification that they hold the right key. So there's no risk of someone just sending / repeating random data without knowing the key.

It's a minor thing anyway since we only truly care about confidentiality from a correctly-running server here and otherwise trust them to not tamper! A "zero-trust sshx" would only be relevant if you're paranoid.

Resolves #13.