Set deadpool config size to 4 conns (#39)
This is a change for server connection pool performance. We only use the Redis connection pool to optimize immediate `GET` and `SET` operations, which should be very fast, so it doesn't make sense to keeping too many idle connections in the pool.

Deadpool's connection pool only configures the number of slots that are maintained for connections to be recycled, so this is fine. Unlike other pools, this will still create more Redis connections if we hit the max pool size.